Let EXPOSED_HOST be container IP by default
On second thought, from writing docs forthcoming, if we're in the
Docker mindset then the container IP ought to be announced by default.
It should be the special case that you use an IP or hostname of the
Docker host, especially 127.0.0.1!

So, `EXPOSED_HOST` should be explicitly set for development if you want
Kafka clients connecting from outside Docker to the interface where you
forward Docker ports. Set it accordingly for 127.0.0.1, or a boot2docker
VM, etc. This should not be onerous since you probably ought to be using
Fig, Vagrant, etc. to do it for you.