Really make JMX usable
The most important piece is specifying the RMI port to match the JMX
port we're exporting -- this prevents random high ports from being
generated on JVM startup which we couldn't deterministically export, so
clients can now connect.

It was also necessary to be able to set the RMI hostname for binding as
well, to be able to connect when the port is forwarded to the Docker
host.